+++ src/raw_data/raw_door.rs
@@ -4,16 +4,16 @@ use crate::utils::Vec3;
 
 #[derive(Debug, Copy, Clone)]
 pub struct RawDoor {
-    pub id: usize,
-    pub wall0_id: usize,
-    wall1_id: usize,
+    pub id: i64,
+    pub wall0_id: i64,
+    wall1_id: i64,
     pub position: Vec3,
     pub width: OrderedFloat<f32>,
     height: OrderedFloat<f32>,
 }
 
 impl RawDoor {
-    pub fn new(id: usize, wall0_id: usize, wall1_id: usize, position: Vec3, width: OrderedFloat<f32>, height: OrderedFloat<f32>) -> Self {
+    pub fn new(id: i64, wall0_id: i64, wall1_id: i64, position: Vec3, width: OrderedFloat<f32>, height: OrderedFloat<f32>) -> Self {
         Self {
             id,
             wall0_id,
